# User/user-api-swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"User Controller","description":"User Management"},{"name":"User Review Controller","description":"User Review Management"},{"name":"Restaurant Recommender Controller","description":"Restaurant Recommendation"}],"paths":{"/api/v1/users/{id}":{"get":{"tags":["User Controller"],"summary":"Get user by user ID","description":"Displays requested user.","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseUserResponse"}}}}}},"put":{"tags":["User Controller"],"summary":"Update User by ID","description":"Users information can update except password.","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseUserResponse"}}}}}},"delete":{"tags":["User Controller"],"summary":"Delete User by ID","description":"The requested user is deleted according to the ID.","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseVoid"}}}}}},"patch":{"tags":["User Controller"],"summary":"Update User Password","description":"The update is done by entering the old and new password.","operationId":"updatePassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdatePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseUserResponse"}}}}}}},"/api/v1/user-reviews/{id}":{"get":{"tags":["User Review Controller"],"summary":"Get User Review by ID","description":"Displays requested User Review.","operationId":"getUserReviewById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseUserReviewResponse"}}}}}},"put":{"tags":["User Review Controller"],"summary":"Update User Review by ID","description":"Score and comment are updated according to the entered ID.","operationId":"updateUserReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReviewUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseUserReviewResponse"}}}}}},"delete":{"tags":["User Review Controller"],"summary":"Delete User Review by ID","description":"The requested User Review is deleted according to the ID.","operationId":"deleteUserReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseVoid"}}}}}}},"/api/v1/users":{"get":{"tags":["User Controller"],"summary":"Get all users","description":"Displays all users.","operationId":"getAllUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseListUserResponse"}}}}}},"post":{"tags":["User Controller"],"summary":"Save User","description":"Save User with necessary information.","operationId":"saveUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseUserResponse"}}}}}}},"/api/v1/user-reviews":{"get":{"tags":["User Review Controller"],"summary":"Get all User Reviews","description":"Displays all User Reviews.","operationId":"getAllUserReviews","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseListUserReviewResponse"}}}}}},"post":{"tags":["User Review Controller"],"summary":"Save User Review","description":"Save User Review with necessary information.","operationId":"saveUserReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReviewSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseUserReviewResponse"}}}}}}},"/api/v1/restaurant-recommenders/{userId}":{"get":{"tags":["Restaurant Recommender Controller"],"summary":"Restaurant Recommendation based on User ID","description":"The best 3 restaurant suggestions are given according to the distance to the location entered by the user and the restaurant score.","operationId":"getTopThreeRestaurants","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RestResponseListRestaurantRecommenderResponse"}}}}}}}},"components":{"schemas":{"UserUpdateRequest":{"required":["email","latitude","longitude","name","status","surname"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string"},"surname":{"maxLength":100,"minLength":0,"type":"string"},"email":{"maxLength":100,"minLength":0,"type":"string"},"status":{"maxLength":10,"minLength":0,"type":"string","enum":["ACTIVE","INACTIVE"]},"birthDate":{"type":"string","format":"date"},"latitude":{"maximum":90.0,"exclusiveMaximum":false,"minimum":-90.0,"exclusiveMinimum":false,"type":"number","format":"double"},"longitude":{"maximum":180.0,"exclusiveMaximum":false,"minimum":-180.0,"exclusiveMinimum":false,"type":"number","format":"double"}}},"RestResponseUserResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/UserResponse"},"responseDate":{"type":"string","format":"date-time"},"messages":{"type":"string"},"success":{"type":"boolean"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"surname":{"type":"string"},"email":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","INACTIVE"]},"birthDate":{"type":"string","format":"date"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"UserReviewUpdateRequest":{"required":["score"],"type":"object","properties":{"score":{"type":"string","enum":["ONE","TWO","THREE","FOUR","FIVE"]},"comment":{"type":"string"}}},"RestResponseUserReviewResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/UserReviewResponse"},"responseDate":{"type":"string","format":"date-time"},"messages":{"type":"string"},"success":{"type":"boolean"}}},"User":{"required":["email","latitude","longitude","name","password","status","surname"],"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"surname":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":30,"minLength":8,"type":"string"},"status":{"type":"string","enum":["ACTIVE","INACTIVE"]},"birthDate":{"type":"string","format":"date"},"latitude":{"maximum":90.0,"exclusiveMaximum":false,"minimum":-90.0,"exclusiveMinimum":false,"type":"number","format":"double"},"longitude":{"maximum":180.0,"exclusiveMaximum":false,"minimum":-180.0,"exclusiveMinimum":false,"type":"number","format":"double"}}},"UserReviewResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"score":{"type":"string","enum":["ONE","TWO","THREE","FOUR","FIVE"]},"comment":{"type":"string"},"user":{"$ref":"#/components/schemas/User"},"restaurantId":{"type":"string"}}},"UserSaveRequest":{"required":["email","latitude","longitude","name","password","surname"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string"},"surname":{"maxLength":100,"minLength":0,"type":"string"},"email":{"maxLength":100,"minLength":0,"type":"string"},"password":{"maxLength":30,"minLength":8,"type":"string"},"birthDate":{"type":"string","format":"date"},"latitude":{"maximum":90.0,"exclusiveMaximum":false,"minimum":-90.0,"exclusiveMinimum":false,"type":"number","format":"double"},"longitude":{"maximum":180.0,"exclusiveMaximum":false,"minimum":-180.0,"exclusiveMinimum":false,"type":"number","format":"double"}}},"UserReviewSaveRequest":{"required":["restaurantId","score","userId"],"type":"object","properties":{"score":{"type":"string","enum":["ONE","TWO","THREE","FOUR","FIVE"]},"comment":{"type":"string"},"userId":{"type":"integer","format":"int64"},"restaurantId":{"type":"string"}}},"UserUpdatePasswordRequest":{"required":["newPassword","newPasswordVerify","oldPassword"],"type":"object","properties":{"oldPassword":{"maxLength":30,"minLength":8,"type":"string"},"newPassword":{"maxLength":30,"minLength":8,"type":"string"},"newPasswordVerify":{"maxLength":30,"minLength":8,"type":"string"}}},"RestResponseListUserResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}},"responseDate":{"type":"string","format":"date-time"},"messages":{"type":"string"},"success":{"type":"boolean"}}},"RestResponseListUserReviewResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserReviewResponse"}},"responseDate":{"type":"string","format":"date-time"},"messages":{"type":"string"},"success":{"type":"boolean"}}},"RestResponseListRestaurantRecommenderResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/RestaurantRecommenderResponse"}},"responseDate":{"type":"string","format":"date-time"},"messages":{"type":"string"},"success":{"type":"boolean"}}},"RestaurantRecommenderResponse":{"type":"object","properties":{"restaurant":{"$ref":"#/components/schemas/RestaurantResponse"},"distance":{"type":"number","format":"double"}}},"RestaurantResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"averageScore":{"type":"number","format":"double"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"RestResponseVoid":{"type":"object","properties":{"data":{"type":"object"},"responseDate":{"type":"string","format":"date-time"},"messages":{"type":"string"},"success":{"type":"boolean"}}}}}}